webscraping/requirements.txt
selenium>=4.10.0
beautifulsoup4>=4.9.0
pandas
openpyxl>=3.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
